♻️ Refactor: 에러처리 리턴
